Cambios en los controllers

+++ src/main/java/com/app/gestiondepedidos/controllers/ClienteController.java
@@ -6,10 +6,11 @@ import com.app.gestiondepedidos.models.Cliente;
 import com.app.gestiondepedidos.services.IClienteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/clientes")
@@ -18,45 +19,70 @@ public class ClienteController {
     @Autowired
     private IClienteService clienteService;
 
-    @GetMapping("")
-    public List<ClienteDTO> index() {
+    @GetMapping
+    public ResponseEntity<List<ClienteDTO>> index() {
         List<Cliente> temp=clienteService.findAll();
-        List<ClienteDTO> lista= new ArrayList<>();
 
-        for(int i = 0; i < temp.size(); i++){
-            ClienteDTO aux= ClienteMapper.crearClienteDTO(temp.get(i));
-            lista.add(aux);
+        if (temp.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
 
-        return lista;
+        List<ClienteDTO> clientes= new ArrayList<>();
+        for(int i = 0; i < temp.size(); i++){
+            ClienteDTO aux= ClienteMapper.crearClienteDTO(temp.get(i));
+            clientes.add(aux);
+        }
+        return new ResponseEntity<>(clientes, HttpStatus.OK);
     }
 
     @GetMapping("{id}")
-    public Cliente show (@PathVariable Long id){
-        return clienteService.findById(id).orElse(null);
+    public ResponseEntity<ClienteDTO> show (@PathVariable Long id){
+        Optional<Cliente> clienteOpt= clienteService.findById(id);
+
+        if (!clienteOpt.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        ClienteDTO cliente= ClienteMapper.crearClienteDTO(clienteOpt.get());
+
+        return new ResponseEntity<>(cliente, HttpStatus.OK);
     }
 
-    @PostMapping("")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Cliente create (@RequestBody Cliente cliente){
-        return clienteService.save(cliente);
+    @PostMapping
+    public ResponseEntity<Cliente> create (@RequestBody Cliente cliente){
+        Cliente clienteCreado= clienteService.save(cliente);
+        return new ResponseEntity<>(clienteCreado, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Cliente update(@RequestBody Cliente cliente, @PathVariable Long id){
-        Cliente clienteActual= clienteService.findById(id).orElse(null);
+    public ResponseEntity<Cliente> update(@RequestBody Cliente cliente, @PathVariable Long id){
+        Optional<Cliente> clienteOpt= clienteService.findById(id);
+
+        if (!clienteOpt.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        Cliente clienteActual= clienteOpt.get();
 
         clienteActual.setNombre(cliente.getNombre());
         clienteActual.setApellido(cliente.getApellido());
         clienteActual.setDni(cliente.getDni());
 
-        return clienteService.save(clienteActual);
+        clienteService.save(clienteActual);
+
+        return new ResponseEntity<>(clienteActual, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete (@PathVariable Long id){
+    public ResponseEntity<Void> delete (@PathVariable Long id){
+        Optional<Cliente> clienteOpt= clienteService.findById(id);
+
+        if (!clienteOpt.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         clienteService.delete(id);
+
+        return  new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
